chore: Refactor Dockerfile and docker-compose for improved development
Fixes an absolute path issue, which in Docker, an absolute path
refers to the full path within the container or on the host. This
is important when mounting volumes, setting working directories,
or copying files.

This change occurred in recent versions of Docker and the first
version of the Dockerfile and Docker Compose files would work
normally in previous versions.